// client/src/store/action/todo/action.ts
export enum CREATE_TODO_ACTION {
    CREATE_TODO_PENDING = "CREATE_TODO_PENDING",
    CREATE_TODO_SUCCESS = "CREATE_TODO_SUCCESS",
    CREATE_TODO_FAILURE = "CREATE_TODO_FAILURE",
}

export enum GET_ALL_TODO_ACTION {
    GET_ALL_TODO_PENDING = "GET_ALL_TODO_PENDING",
    GET_ALL_TODO_SUCCESS = "GET_ALL_TODO_SUCCESS",
    GET_ALL_TODO_FAILURE = "GET_ALL_TODO_FAILURE",
}

export enum ADMIN_GET_ALL_TODO_ACTION {
    ADMIN_GET_ALL_TODO_PENDING = "ADMIN_GET_ALL_TODO_PENDING",
    ADMIN_GET_ALL_TODO_SUCCESS = "ADMIN_GET_ALL_TODO_SUCCESS",
    ADMIN_GET_ALL_TODO_FAILURE = "ADMIN_GET_ALL_TODO_FAILURE",
}

export enum UPDATE_TODO_ACTION {
    UPDATE_TODO_PENDING = "UPDATE_TODO_PENDING",
    UPDATE_TODO_SUCCESS = "UPDATE_TODO_SUCCESS",
    UPDATE_TODO_FAILURE = "UPDATE_TODO_FAILURE",
}

export enum DELETE_TODO_ACTION {
    DELETE_TODO_PENDING = "DELETE_TODO_PENDING",
    DELETE_TODO_SUCCESS = "DELETE_TODO_SUCCESS",
    DELETE_TODO_FAILURE = "DELETE_TODO_FAILURE",
}

export enum GET_SINGLE_TODO_ACTION {
    GET_SINGLE_TODO_PENDING = "GET_SINGLE_TODO_PENDING",
    GET_SINGLE_TODO_SUCCESS = "GET_SINGLE_TODO_SUCCESS",
    GET_SINGLE_TODO_FAILURE = "GET_SINGLE_TODO_FAILURE",
}

export enum ADMIN_GET_SINGLE_TODO_ACTION {
    ADMIN_GET_SINGLE_TODO_PENDING = "ADMIN_GET_SINGLE_TODO_PENDING",
    ADMIN_GET_SINGLE_TODO_SUCCESS = "ADMIN_GET_SINGLE_TODO_SUCCESS",
    ADMIN_GET_SINGLE_TODO_FAILURE = "ADMIN_GET_SINGLE_TODO_FAILURE",
}